requirements.txt
flask
fpdf
pillow